Predicting the nature of movie review sentiments
Datasets: From a reputed movie review and rating website

The model used in this study is based on the infamous Naive Bayes theorem which is a simple-yet-powerful classification algorithm. The ROC-AUC measure was also calculated to compute the error between the actual values and the predicted values in the test dataset.